+++ src/short-urls/helpers/ShortUrlsFilterDropdown.tsx
@@ -16,7 +16,7 @@ export const ShortUrlsFilterDropdown = (
   const onFilterClick = (key: keyof ShortUrlsFilter) => () => onChange({ ...selected, [key]: !selected?.[key] });
 
   return (
-    <DropdownBtn text="Filters" dropdownClassName={className} inline end minWidth={250}>
+    <DropdownBtn text="Filters" dropdownClassName={className} end minWidth={250}>
       <DropdownItem header aria-hidden>Visits:</DropdownItem>
       <DropdownItem active={excludeBots} onClick={onFilterClick('excludeBots')}>Ignore visits from bots</DropdownItem>