+++ src/utils/dates/DateRangeSelector.tsx
@@ -50,14 +50,13 @@ export const DateRangeSelector = (
     <DropdownBtn disabled={disabled} text={text}>
       <DateIntervalDropdownItems allText={defaultText} active={activeInterval} onChange={updateDateRangeOrInterval} />
       <DropdownItem divider tag="hr" />
-      <DropdownItem header aria-hidden>Custom:</DropdownItem>
-      <DropdownItem text>
+      <div className="px-3 py-1">
         <DateRangeRow
           {...activeDateRange}
           onStartDateChange={(startDate) => updateDateRangeOrInterval({ ...activeDateRange, startDate })}
           onEndDateChange={(endDate) => updateDateRangeOrInterval({ ...activeDateRange, endDate })}
         />
-      </DropdownItem>
+      </div>
     </DropdownBtn>
   );
 };